fix(P0): fix CI-blocking defects from ROADMAP_PRODUCTION.md P0.1-P0.5
- P0.1: Change organization_id/device_id/acknowledged_by INT UNSIGNED→BIGINT
  UNSIGNED in migration 163 and schema.sql to fix ERROR 3780 FK mismatch
- P0.2: Add splitStatements() DELIMITER-aware parser to migrate.js so
  migration 028 stored procedures execute correctly via mysql2; 7 new tests
  in tests/migrate.test.js
- P0.3: Inline safe integer limit/offset directly into BaseModel.findAll SQL
  string (not prepared-statement params) to fix mysqld_stmt_execute regression
  on all paginated list endpoints; update crudController test
- P0.4: Fix README.md migration range (001-163) and table count (110); flip
  database-tests WARNING to hard exit 1
- P0.5: Add FK column type CI assertion using information_schema

Agent-Logs-Url: https://github.com/vothalvino/fireisp5.0/sessions/2b8c7abf-be88-4734-903e-d41c9f75230b

+++ database/schema.sql
@@ -1108,8 +1108,8 @@ CREATE TABLE IF NOT EXISTS snmp_profile_oids (
 
 CREATE TABLE IF NOT EXISTS snmp_traps (
     id               BIGINT UNSIGNED  NOT NULL AUTO_INCREMENT,
-    organization_id  INT UNSIGNED,
-    device_id        INT UNSIGNED,
+    organization_id  BIGINT UNSIGNED,
+    device_id        BIGINT UNSIGNED,
     source_ip        VARCHAR(45)      NOT NULL,
     trap_type        VARCHAR(64)      NOT NULL DEFAULT 'unknown',
     trap_oid         VARCHAR(255),
@@ -1117,7 +1117,7 @@ CREATE TABLE IF NOT EXISTS snmp_traps (
     community        VARCHAR(128),
     snmp_version     TINYINT UNSIGNED NOT NULL DEFAULT 2,
     is_acknowledged  TINYINT(1)       NOT NULL DEFAULT 0,
-    acknowledged_by  INT UNSIGNED,
+    acknowledged_by  BIGINT UNSIGNED,
     acknowledged_at  DATETIME,
     received_at      DATETIME         NOT NULL DEFAULT CURRENT_TIMESTAMP,
